+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/CommonAbstracts/types/InlineMessageAction.kt
@@ -1,0 +1,7 @@
+package com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.InlineMessageIdentifier
+
+interface InlineMessageAction {
+    val inlineMessageId: InlineMessageIdentifier
+}