+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/CommonAbstracts/types/MessageAction.kt
@@ -1,0 +1,7 @@
+package com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types
+
+import com.github.insanusmokrassar.TelegramBotAPI.types.MessageIdentifier
+
+interface MessageAction: ChatRequest {
+    val messageId: MessageIdentifier
+}