+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/DeleteMessage.kt
@@ -1,5 +1,6 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.MessageAction
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import kotlinx.serialization.*
@@ -8,10 +9,10 @@ import kotlinx.serialization.internal.BooleanSerializer
 @Serializable
 data class DeleteMessage(
     @SerialName(chatIdField)
-    val chatId: ChatIdentifier,
+    override val chatId: ChatIdentifier,
     @SerialName(messageIdField)
-    val messageId: MessageIdentifier
-) : SimpleRequest<Boolean> {
+    override val messageId: MessageIdentifier
+) : SimpleRequest<Boolean>, MessageAction {
     override fun method(): String = "deleteMessage"
 
     override fun resultSerializer(): KSerializer<Boolean> = BooleanSerializer