+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/ForwardMessage.kt
@@ -1,5 +1,6 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests
 
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.MessageAction
 import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import com.github.insanusmokrassar.TelegramBotAPI.types.message.RawMessage
@@ -12,10 +13,14 @@ data class ForwardMessage(
     @SerialName(chatIdField)
     val toChatId: ChatIdentifier,
     @SerialName(messageIdField)
-    val messageId: MessageIdentifier,
+    override val messageId: MessageIdentifier,
     @SerialName(disableNotificationField)
     val disableNotification: Boolean = false
-): SimpleRequest<RawMessage> {
+): SimpleRequest<RawMessage>, MessageAction {
+    @Transient
+    override val chatId: ChatIdentifier
+        get() = fromChatId
+
     override fun method(): String = "forwardMessage"
 
     override fun resultSerializer(): KSerializer<RawMessage> = RawMessage.serializer()