+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/StopPoll.kt
@@ -1,0 +1,24 @@
+package com.github.insanusmokrassar.TelegramBotAPI.requests
+
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.*
+import com.github.insanusmokrassar.TelegramBotAPI.requests.abstracts.SimpleRequest
+import com.github.insanusmokrassar.TelegramBotAPI.requests.send.abstracts.SendChatMessageRequest
+import com.github.insanusmokrassar.TelegramBotAPI.requests.send.abstracts.SendMessageRequest
+import com.github.insanusmokrassar.TelegramBotAPI.types.*
+import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.InlineKeyboardMarkup
+import com.github.insanusmokrassar.TelegramBotAPI.types.buttons.KeyboardMarkup
+import com.github.insanusmokrassar.TelegramBotAPI.types.polls.Poll
+import kotlinx.serialization.*
+
+@Serializable
+data class StopPoll(
+    @SerialName(chatIdField)
+    override val chatId: ChatIdentifier,
+    @SerialName(messageIdField)
+    override val messageId: MessageIdentifier,
+    @SerialName(replyMarkupField)
+    override val replyMarkup: InlineKeyboardMarkup?
+) : MessageAction, SimpleRequest<Poll>, ReplyMarkup {
+    override fun method(): String = "stopPoll"
+    override fun resultSerializer(): KSerializer<Poll> = Poll.serializer()
+}