+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/games/GetGameHighScoresByInlineMessageId.kt
@@ -1,6 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests.games
 
 import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.ByInlineMessageId
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.InlineMessageAction
 import com.github.insanusmokrassar.TelegramBotAPI.requests.games.abstracts.GetGameHighScores
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import kotlinx.serialization.SerialName
@@ -12,4 +13,4 @@ data class GetGameHighScoresByInlineMessageId (
     override val userId: UserId,
     @SerialName(inlineMessageIdField)
     override val inlineMessageId: InlineMessageIdentifier
-) : GetGameHighScores, ByInlineMessageId
+) : GetGameHighScores, InlineMessageAction