+++ src/main/kotlin/com/github/insanusmokrassar/TelegramBotAPI/requests/games/SetGameScoreByInlineMessageId.kt
@@ -1,6 +1,7 @@
 package com.github.insanusmokrassar.TelegramBotAPI.requests.games
 
 import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.ByInlineMessageId
+import com.github.insanusmokrassar.TelegramBotAPI.CommonAbstracts.types.InlineMessageAction
 import com.github.insanusmokrassar.TelegramBotAPI.requests.games.abstracts.SetGameScore
 import com.github.insanusmokrassar.TelegramBotAPI.types.*
 import kotlinx.serialization.SerialName
@@ -18,4 +19,4 @@ data class SetGameScoreByInlineMessageId (
     override val force: Boolean = false,
     @SerialName(disableEditMessageField)
     override val disableEditMessage: Boolean = false
-) : SetGameScore, ByInlineMessageId
+) : SetGameScore, InlineMessageAction